refactor: rename resource API methods for clarity and consistency
BREAKING CHANGE: Resource API method names have been updated

Thin-flat API (ResourceClient):
- list() → fetchList() (explicit network action)
- get() → fetch() (explicit network action)

Grouped API (ResourceObject):
- fetch.item() → get.item() (general intent)
- fetch.list() → get.list() (general intent)

Rationale:
- Thin-flat methods use 'fetch' prefix to clearly indicate network calls
- Grouped API uses 'get' namespace for semantic clarity
- Avoids confusion with global fetch() API
- Eliminates property name collisions between APIs

Migration guide:
```typescript
// OLD thin-flat API:
await testimonial.list(query)
await testimonial.get(id)

// NEW thin-flat API:
await testimonial.fetchList(query)
await testimonial.fetch(id)

// OLD grouped API:
await testimonial.fetch.item(id)
await testimonial.fetch.list(query)

// NEW grouped API:
await testimonial.get.item(id)
await testimonial.get.list(query)
```

Changes:
- Updated type definitions (ResourceClient, ResourceObject)
- Updated all implementations (client.ts, grouped-api.ts, define.ts, store.ts)
- Fixed all resolvers to check for new method names
- Updated 383 unit tests (all passing)
- Updated all documentation guides
- Added comprehensive resources-advanced.md guide
- Reorganized test files for better maintainability

Tests: 383/383 passing ✅

Closes #28
