+++ docs/guide/actions.md
@@ -4,24 +4,45 @@
 
 Actions orchestrate writes. They:
 
-- Call Resources
-- Emit Events
-- Invalidate cache
-- Enqueue Jobs
+- Call Resources (write operations)
+- Emit Events (for extensibility)
+- Invalidate cache (keep UI fresh)
+- Enqueue Jobs (background work)
+
+**Golden Rule**: UI components **never** call resource write methods directly. Always route through Actions.
 
 ## Quick Reference
 
 ```typescript
 import { defineAction } from '@geekist/wp-kernel/actions';
+import { testimonial } from '@/resources/Testimonial';
 import { events } from '@geekist/wp-kernel/events';
-import { invalidate } from '@wordpress/data';
 
-export const CreateThing = defineAction(
-	'Thing.Create',
-	async ({ data }: { data: Partial<Thing> }) => {
-		const created = await thing.create(data);
-		CreateThing.emit(events.thing.created, { id: created.id, data });
-		invalidate(['thing', 'list']);
+export const CreateTestimonial = defineAction(
+	'Testimonial.Create',
+	async ({ data }: { data: Partial<TestimonialPost> }) => {
+		// Permission check
+		if (!currentUserCan('create_testimonials')) {
+			throw new PolicyDenied('testimonials.create');
+		}
+
+		// Call resource
+		const created = await testimonial.create(data);
+
+		// Emit event
+		CreateTestimonial.emit(events.testimonial.created, {
+			id: created.id,
+			data: created,
+		});
+
+		// Invalidate cache
+		testimonial.invalidate([['testimonial', 'list']]);
+
+		// Queue job if needed
+		if (data.featured) {
+			await jobs.enqueue('NotifyFeaturedTestimonial', { id: created.id });
+		}
+
 		return created;
 	}
 );
@@ -29,5 +50,7 @@ export const CreateThing = defineAction(
 
 ## See Also
 
+- [Resources Guide](/guide/resources) - Using resources in actions
+- [Events Guide](/guide/events) - Event emission patterns
 - [Quick Start](/getting-started/quick-start) - Build your first action
 - [API Reference](/api/actions) - Complete API docs (coming soon)
