+++ docs/guide/block-bindings.md
@@ -2,7 +2,7 @@
 
 > **Status**: 🚧 This page will be expanded in Sprint 1+
 
-Bind core WordPress blocks to your store data.
+Bind core WordPress blocks to your store data. No custom blocks needed for read paths.
 
 ## Quick Reference
 
@@ -11,8 +11,12 @@ import { registerBindingSource } from '@geekist/wp-kernel/bindings';
 import { select } from '@wordpress/data';
 
 registerBindingSource('gk', {
-	'thing.title': (attrs) => select('wpk/thing').getById(attrs.id)?.title,
-	'thing.price': (attrs) => select('wpk/thing').getById(attrs.id)?.price,
+	'testimonial.title': (attrs) =>
+		select('wpk/testimonial').getById(attrs.id)?.title,
+	'testimonial.content': (attrs) =>
+		select('wpk/testimonial').getById(attrs.id)?.content,
+	'testimonial.rating': (attrs) =>
+		select('wpk/testimonial').getById(attrs.id)?.rating,
 });
 ```
 
@@ -21,12 +25,13 @@ In `block.json`:
 ```json
 {
 	"bindings": {
-		"core/heading": { "content": "gk:thing.title" },
-		"core/paragraph": { "content": "gk:thing.price" }
+		"core/heading": { "content": "gk:testimonial.title" },
+		"core/paragraph": { "content": "gk:testimonial.content" }
 	}
 }
 ```
 
 ## See Also
 
+- [Resources Guide](/guide/resources) - Using resources with bindings
 - [WordPress Block Bindings API](https://developer.wordpress.org/block-editor/reference-guides/block-api/block-bindings/)
