+++ packages/kernel/src/index.ts
@@ -1,19 +1,47 @@
 /**
  * WP Kernel - Core Framework Package
  *
- * Rails-like framework for building modern WordPress products
+ * Rails-like framework for building modern WordPress products where
+ * JavaScript is the source of truth and PHP is a thin contract.
+ *
+ * @example Scoped imports (recommended)
+ * ```ts
+ * import { fetch } from '@geekist/wp-kernel/http';
+ * import { defineResource } from '@geekist/wp-kernel/resource';
+ * import { KernelError } from '@geekist/wp-kernel/error';
+ * ```
+ *
+ * @example Namespace imports (organized)
+ * ```ts
+ * import { http, resource, error } from '@geekist/wp-kernel';
+ * await http.fetch({ path: '/wpk/v1/things' });
+ * resource.defineResource({ name: 'thing', routes: {...} });
+ * throw new error.KernelError('ValidationError', {...});
+ * ```
+ *
+ * @example Flat imports (convenience)
+ * ```ts
+ * import { fetch, defineResource, KernelError } from '@geekist/wp-kernel';
+ * ```
  *
  * @module
  */
 
-/**
- * Current version of WP Kernel
- */
-export const VERSION = '0.0.0';
+export const VERSION = '0.1.1';
 
-/**
- * Error types (Sprint 1)
- */
+// ============================================================================
+// Namespace Exports (Organized by module)
+// ============================================================================
+
+export * as http from './http/index.js';
+export * as resource from './resource/index.js';
+export * as error from './error/index.js';
+
+// ============================================================================
+// Flat Exports (Convenience aliases)
+// ============================================================================
+
+// Error classes
 export { KernelError, TransportError, ServerError } from './error/index.js';
 export type {
 	ErrorCode,
@@ -22,9 +50,7 @@ export type {
 	SerializedError,
 } from './error/index.js';
 
-/**
- * Transport layer (Sprint 1 - A3)
- */
+// HTTP transport
 export { fetch } from './http/fetch.js';
 export type {
 	HttpMethod,
@@ -35,11 +61,9 @@ export type {
 	ResourceErrorEvent,
 } from './http/types.js';
 
-/**
- * Resource system (Sprint 1)
- */
+// Resource system
+export { defineResource } from './resource/define.js';
 export {
-	defineResource,
 	interpolatePath,
 	extractPathParams,
 	invalidate,
@@ -48,7 +72,7 @@ export {
 	matchesCacheKey,
 	findMatchingKeys,
 	findMatchingKeysMultiple,
-} from './resource/index.js';
+} from './resource/cache.js';
 export { createStore } from './resource/store.js';
 export type {
 	ResourceRoute,
@@ -59,11 +83,6 @@ export type {
 	ListResponse,
 	ResourceClient,
 	ResourceObject,
-	PathParams,
-	CacheKeyPattern,
-	InvalidateOptions,
-} from './resource/index.js';
-export type {
 	ResourceState,
 	ResourceActions,
 	ResourceSelectors,
@@ -71,25 +90,8 @@ export type {
 	ResourceStoreConfig,
 	ResourceStore,
 } from './resource/types.js';
-
-/**
- * Module placeholders - implementations coming in future sprints
- *
- * Planned exports:
- * - defineAction (from './actions.js') - Sprint 3
- * - definePolicy (from './policies.js') - Sprint 2
- * - defineJob (from './jobs.js') - Sprint 4
- * - defineInteraction (from './interactivity.js') - Sprint 5
- * - registerBindingSource (from './bindings.js') - Sprint 6
- * - events (from './events.js') - Sprint 1
- */
-
-// Future module structure:
-// export * from './resource/index.js';
-// export * from './actions.js';
-// export * from './policies.js';
-// export * from './jobs.js';
-// export * from './interactivity.js';
-// export * from './bindings.js';
-// export * from './events.js';
-// export * from './errors/index.js';
+export type {
+	PathParams,
+	CacheKeyPattern,
+	InvalidateOptions,
+} from './resource/cache.js';
