+++ packages/kernel/src/resource/__tests__/store.test.ts
@@ -51,9 +51,12 @@ describe('createStore', () => {
 			routes: {
 				list: { path: '/wpk/v1/things', method: 'GET' },
 				get: { path: '/wpk/v1/things/:id', method: 'GET' },
+				create: { path: '/wpk/v1/things', method: 'POST' },
+				update: { path: '/wpk/v1/things/:id', method: 'PUT' },
+				remove: { path: '/wpk/v1/things/:id', method: 'DELETE' },
 			},
-			list: jest.fn().mockResolvedValue(mockListResponse),
-			get: jest.fn().mockResolvedValue({
+			fetchList: jest.fn().mockResolvedValue(mockListResponse),
+			fetch: jest.fn().mockResolvedValue({
 				id: 1,
 				title: 'Thing One',
 				status: 'active',
@@ -89,6 +92,58 @@ describe('createStore', () => {
 				}
 			),
 			store: {},
+			// Grouped API namespaces
+			select: {
+				item: jest.fn().mockReturnValue(undefined),
+				items: jest.fn().mockReturnValue([]),
+				list: jest.fn().mockReturnValue([]),
+			},
+			use: {
+				item: jest.fn(),
+				list: jest.fn(),
+			},
+			get: {
+				item: jest.fn().mockResolvedValue({
+					id: 1,
+					title: 'Thing One',
+					status: 'active',
+				}),
+				list: jest.fn().mockResolvedValue(mockListResponse),
+			},
+			mutate: {
+				create: jest.fn().mockResolvedValue({
+					id: 3,
+					title: 'New Thing',
+					status: 'active',
+				}),
+				update: jest.fn().mockResolvedValue({
+					id: 1,
+					title: 'Updated Thing',
+					status: 'active',
+				}),
+				remove: jest.fn().mockResolvedValue(undefined),
+			},
+			cache: {
+				prefetch: {
+					item: jest.fn().mockResolvedValue(undefined),
+					list: jest.fn().mockResolvedValue(undefined),
+				},
+				invalidate: {
+					item: jest.fn(),
+					list: jest.fn(),
+					all: jest.fn(),
+				},
+				key: jest.fn(),
+			},
+			storeApi: {
+				key: 'wpk/thing',
+				descriptor: {},
+			},
+			events: {
+				created: 'wpk.thing.created',
+				updated: 'wpk.thing.updated',
+				removed: 'wpk.thing.removed',
+			},
 		};
 	});
 
@@ -528,11 +583,11 @@ describe('createStore', () => {
 		describe('getItem', () => {
 			it('should fetch item and return receiveItem action', async () => {
 				const item = { id: 1, title: 'Thing One', status: 'active' };
-				(mockResource.get as jest.Mock).mockResolvedValue(item);
+				(mockResource.fetch as jest.Mock).mockResolvedValue(item);
 
 				const action = await store.resolvers.getItem(1);
 
-				expect(mockResource.get).toHaveBeenCalledWith(1);
+				expect(mockResource.fetch).toHaveBeenCalledWith(1);
 				expect(action).toEqual({
 					type: 'RECEIVE_ITEM',
 					item,
@@ -541,7 +596,7 @@ describe('createStore', () => {
 
 			it('should return receiveError action on fetch failure', async () => {
 				const error = new Error('Network error');
-				(mockResource.get as jest.Mock).mockRejectedValue(error);
+				(mockResource.fetch as jest.Mock).mockRejectedValue(error);
 
 				const action = await store.resolvers.getItem(1);
 
@@ -552,35 +607,35 @@ describe('createStore', () => {
 				});
 			});
 
-			it('should throw NotImplementedError when get method not defined', async () => {
-				const resourceWithoutGet = {
+			it('should throw NotImplementedError when fetch method not defined', async () => {
+				const resourceWithoutFetch = {
 					...mockResource,
-					get: undefined,
+					fetch: undefined,
 				};
 
-				const storeWithoutGet = createStore({
-					resource: resourceWithoutGet,
+				const storeWithoutFetch = createStore({
+					resource: resourceWithoutFetch,
 				});
 
 				await expect(
-					storeWithoutGet.resolvers.getItem(1)
+					storeWithoutFetch.resolvers.getItem(1)
 				).rejects.toThrow(KernelError);
 				await expect(
-					storeWithoutGet.resolvers.getItem(1)
-				).rejects.toThrow(/does not have a "get" method/);
+					storeWithoutFetch.resolvers.getItem(1)
+				).rejects.toThrow(/does not have a "fetch" method/);
 			});
 		});
 
 		describe('getItems', () => {
 			it('should fetch items and return receiveItems action', async () => {
 				const query = { q: 'search' };
-				(mockResource.list as jest.Mock).mockResolvedValue(
+				(mockResource.fetchList as jest.Mock).mockResolvedValue(
 					mockListResponse
 				);
 
 				const action = await store.resolvers.getItems(query);
 
-				expect(mockResource.list).toHaveBeenCalledWith(query);
+				expect(mockResource.fetchList).toHaveBeenCalledWith(query);
 				expect(action).toEqual({
 					type: 'RECEIVE_ITEMS',
 					queryKey: JSON.stringify(query),
@@ -594,20 +649,20 @@ describe('createStore', () => {
 			});
 
 			it('should handle query without parameters', async () => {
-				(mockResource.list as jest.Mock).mockResolvedValue(
+				(mockResource.fetchList as jest.Mock).mockResolvedValue(
 					mockListResponse
 				);
 
 				const action = await store.resolvers.getItems();
 
-				expect(mockResource.list).toHaveBeenCalledWith(undefined);
+				expect(mockResource.fetchList).toHaveBeenCalledWith(undefined);
 				expect(action).toHaveProperty('type', 'RECEIVE_ITEMS');
 				expect(action).toHaveProperty('queryKey');
 			});
 
 			it('should return receiveError action on fetch failure', async () => {
 				const error = new Error('Network error');
-				(mockResource.list as jest.Mock).mockRejectedValue(error);
+				(mockResource.fetchList as jest.Mock).mockRejectedValue(error);
 
 				const action = await store.resolvers.getItems();
 
@@ -618,10 +673,10 @@ describe('createStore', () => {
 				});
 			});
 
-			it('should throw NotImplementedError when list method not defined', async () => {
+			it('should throw NotImplementedError when fetchList method not defined', async () => {
 				const resourceWithoutList = {
 					...mockResource,
-					list: undefined,
+					fetchList: undefined,
 				};
 
 				const storeWithoutList = createStore({
@@ -633,20 +688,20 @@ describe('createStore', () => {
 				).rejects.toThrow(KernelError);
 				await expect(
 					storeWithoutList.resolvers.getItems()
-				).rejects.toThrow(/does not have a "list" method/);
+				).rejects.toThrow(/does not have a "fetchList" method/);
 			});
 		});
 
 		describe('getList', () => {
 			it('should delegate to getItems resolver', async () => {
 				const query = { status: 'active' };
-				(mockResource.list as jest.Mock).mockResolvedValue(
+				(mockResource.fetchList as jest.Mock).mockResolvedValue(
 					mockListResponse
 				);
 
 				const action = await store.resolvers.getList(query);
 
-				expect(mockResource.list).toHaveBeenCalledWith(query);
+				expect(mockResource.fetchList).toHaveBeenCalledWith(query);
 				expect(action).toEqual({
 					type: 'RECEIVE_ITEMS',
 					queryKey: JSON.stringify(query),
@@ -932,7 +987,7 @@ describe('createStore', () => {
 		describe('integration with store', () => {
 			it('should work together - fetch, key, invalidate', async () => {
 				// Fetch an item
-				const item = await mockResource.get?.(123);
+				const item = await mockResource.fetch?.(123);
 				expect(item).toBeDefined();
 
 				// Generate cache key for the item
@@ -961,7 +1016,7 @@ describe('createStore', () => {
 				]);
 
 				// Fetch the list
-				const result = await mockResource.list?.(query);
+				const result = await mockResource.fetchList?.(query);
 				expect(result?.items).toHaveLength(2);
 			});
 
@@ -1014,17 +1069,16 @@ describe('createStore', () => {
 							method: 'GET' as const,
 						},
 					},
-					list: undefined,
+					fetchList: undefined,
 					useList: undefined,
 					prefetchList: undefined,
 				};
 
-				expect(getOnlyResource.get).toBeDefined();
+				expect(getOnlyResource.fetch).toBeDefined();
 				expect(getOnlyResource.useGet).toBeDefined();
 				expect(getOnlyResource.prefetchGet).toBeDefined();
-				expect(getOnlyResource.list).toBeUndefined();
+				expect(getOnlyResource.fetchList).toBeUndefined();
 			});
-
 			it('should work with only list route defined', () => {
 				const listOnlyResource = {
 					...mockResource,
@@ -1034,15 +1088,415 @@ describe('createStore', () => {
 							method: 'GET' as const,
 						},
 					},
-					get: undefined,
+					fetch: undefined,
 					useGet: undefined,
 					prefetchGet: undefined,
 				};
 
-				expect(listOnlyResource.list).toBeDefined();
+				expect(listOnlyResource.fetchList).toBeDefined();
 				expect(listOnlyResource.useList).toBeDefined();
 				expect(listOnlyResource.prefetchList).toBeDefined();
-				expect(listOnlyResource.get).toBeUndefined();
+				expect(listOnlyResource.fetch).toBeUndefined();
+			});
+		});
+	});
+	describe('grouped API namespaces', () => {
+		describe('select namespace (pure selectors)', () => {
+			it('should have select.item method', () => {
+				expect(mockResource.select).toBeDefined();
+				expect(mockResource.select?.item).toBeDefined();
+				expect(typeof mockResource.select?.item).toBe('function');
+			});
+
+			it('should have select.items method', () => {
+				expect(mockResource.select?.items).toBeDefined();
+				expect(typeof mockResource.select?.items).toBe('function');
+			});
+
+			it('should have select.list method', () => {
+				expect(mockResource.select?.list).toBeDefined();
+				expect(typeof mockResource.select?.list).toBe('function');
+			});
+
+			it('select.item should return undefined for non-existent items', () => {
+				const item = mockResource.select?.item(123);
+				expect(item).toBeUndefined();
+			});
+
+			it('select.items should return empty array when no items', () => {
+				const items = mockResource.select?.items();
+				expect(items).toEqual([]);
+			});
+
+			it('select.list should return empty array when no list', () => {
+				const list = mockResource.select?.list({ q: 'search' });
+				expect(list).toEqual([]);
+			});
+		});
+
+		describe('use namespace (React hooks)', () => {
+			it('should have use.item hook', () => {
+				expect(mockResource.use).toBeDefined();
+				expect(mockResource.use?.item).toBeDefined();
+				expect(typeof mockResource.use?.item).toBe('function');
+			});
+
+			it('should have use.list hook', () => {
+				expect(mockResource.use?.list).toBeDefined();
+				expect(typeof mockResource.use?.list).toBe('function');
+			});
+
+			it('use.item should be callable', () => {
+				expect(() => mockResource.use?.item(123)).not.toThrow();
+			});
+
+			it('use.list should be callable with query', () => {
+				expect(() =>
+					mockResource.use?.list({ q: 'search' })
+				).not.toThrow();
+			});
+
+			it('use.list should be callable without query', () => {
+				expect(() => mockResource.use?.list()).not.toThrow();
+			});
+		});
+
+		describe('fetch namespace (explicit fetching)', () => {
+			it('should have fetch.item method', () => {
+				expect(mockResource.fetch).toBeDefined();
+				expect(mockResource.get?.item).toBeDefined();
+				expect(typeof mockResource.get?.item).toBe('function');
+			});
+
+			it('should have fetch.list method', () => {
+				expect(mockResource.get?.list).toBeDefined();
+				expect(typeof mockResource.get?.list).toBe('function');
+			});
+
+			it('fetch.item should fetch single item', async () => {
+				const item = await mockResource.get?.item(123);
+				expect(item).toEqual({
+					id: 1,
+					title: 'Thing One',
+					status: 'active',
+				});
+			});
+
+			it('fetch.list should fetch list of items', async () => {
+				const result = await mockResource.get?.list({ q: 'search' });
+				expect(result).toEqual(mockListResponse);
+			});
+		});
+
+		describe('mutate namespace (CRUD operations)', () => {
+			it('should have mutate.create method', () => {
+				expect(mockResource.mutate).toBeDefined();
+				expect(mockResource.mutate?.create).toBeDefined();
+				expect(typeof mockResource.mutate?.create).toBe('function');
+			});
+
+			it('should have mutate.update method', () => {
+				expect(mockResource.mutate?.update).toBeDefined();
+				expect(typeof mockResource.mutate?.update).toBe('function');
+			});
+
+			it('should have mutate.remove method', () => {
+				expect(mockResource.mutate?.remove).toBeDefined();
+				expect(typeof mockResource.mutate?.remove).toBe('function');
+			});
+
+			it('mutate.create should create new item', async () => {
+				const data = { title: 'New Thing', status: 'active' };
+				const result = await mockResource.mutate?.create(data);
+				expect(result).toEqual({
+					id: 3,
+					title: 'New Thing',
+					status: 'active',
+				});
+			});
+
+			it('mutate.update should update existing item', async () => {
+				const data = { title: 'Updated Thing' };
+				const result = await mockResource.mutate?.update(1, data);
+				expect(result).toEqual({
+					id: 1,
+					title: 'Updated Thing',
+					status: 'active',
+				});
+			});
+
+			it('mutate.remove should delete item', async () => {
+				const result = await mockResource.mutate?.remove(1);
+				expect(result).toBeUndefined();
+			});
+		});
+
+		describe('cache namespace (cache control)', () => {
+			it('should have cache.prefetch.item method', () => {
+				expect(mockResource.cache).toBeDefined();
+				expect(mockResource.cache.prefetch.item).toBeDefined();
+				expect(typeof mockResource.cache.prefetch.item).toBe(
+					'function'
+				);
+			});
+
+			it('should have cache.prefetch.list method', () => {
+				expect(mockResource.cache.prefetch.list).toBeDefined();
+				expect(typeof mockResource.cache.prefetch.list).toBe(
+					'function'
+				);
+			});
+
+			it('should have cache.invalidate.item method', () => {
+				expect(mockResource.cache.invalidate.item).toBeDefined();
+				expect(typeof mockResource.cache.invalidate.item).toBe(
+					'function'
+				);
+			});
+
+			it('should have cache.invalidate.list method', () => {
+				expect(mockResource.cache.invalidate.list).toBeDefined();
+				expect(typeof mockResource.cache.invalidate.list).toBe(
+					'function'
+				);
+			});
+
+			it('should have cache.invalidate.all method', () => {
+				expect(mockResource.cache.invalidate.all).toBeDefined();
+				expect(typeof mockResource.cache.invalidate.all).toBe(
+					'function'
+				);
+			});
+
+			it('should have cache.key property', () => {
+				expect(mockResource.cache.key).toBeDefined();
+			});
+
+			it('cache.prefetch.item should prefetch single item', async () => {
+				await mockResource.cache.prefetch.item(123);
+				expect(mockResource.cache.prefetch.item).toHaveBeenCalledWith(
+					123
+				);
+			});
+
+			it('cache.prefetch.list should prefetch list', async () => {
+				const query = { q: 'search' };
+				await mockResource.cache.prefetch.list(query);
+				expect(mockResource.cache.prefetch.list).toHaveBeenCalledWith(
+					query
+				);
+			});
+
+			it('cache.invalidate.item should invalidate single item', () => {
+				mockResource.cache.invalidate.item(123);
+				expect(mockResource.cache.invalidate.item).toHaveBeenCalledWith(
+					123
+				);
+			});
+
+			it('cache.invalidate.list should invalidate list', () => {
+				const query = { q: 'search' };
+				mockResource.cache.invalidate.list(query);
+				expect(mockResource.cache.invalidate.list).toHaveBeenCalledWith(
+					query
+				);
+			});
+
+			it('cache.invalidate.all should invalidate all', () => {
+				mockResource.cache.invalidate.all();
+				expect(
+					mockResource.cache.invalidate.all
+				).toHaveBeenCalledWith();
+			});
+		});
+
+		describe('storeApi namespace (store access)', () => {
+			it('should have storeApi.key property', () => {
+				expect(mockResource.storeApi).toBeDefined();
+				expect(mockResource.storeApi?.key).toBe('wpk/thing');
+			});
+
+			it('should have storeApi.descriptor property', () => {
+				expect(mockResource.storeApi?.descriptor).toBeDefined();
+				expect(typeof mockResource.storeApi?.descriptor).toBe('object');
+			});
+		});
+
+		describe('events namespace (event names)', () => {
+			it('should have events.created property', () => {
+				expect(mockResource.events).toBeDefined();
+				expect(mockResource.events?.created).toBe('wpk.thing.created');
+			});
+
+			it('should have events.updated property', () => {
+				expect(mockResource.events?.updated).toBe('wpk.thing.updated');
+			});
+
+			it('should have events.removed property', () => {
+				expect(mockResource.events?.removed).toBe('wpk.thing.removed');
+			});
+
+			it('should follow wpk.{name}.{action} naming pattern', () => {
+				expect(mockResource.events?.created).toMatch(/^wpk\.\w+\.\w+$/);
+				expect(mockResource.events?.updated).toMatch(/^wpk\.\w+\.\w+$/);
+				expect(mockResource.events?.removed).toMatch(/^wpk\.\w+\.\w+$/);
+			});
+		});
+
+		describe('grouped API integration patterns', () => {
+			it('should support read path: select → use → fetch', async () => {
+				// Pure selector (no fetch)
+				const cachedItem = mockResource.select?.item(123);
+				expect(cachedItem).toBeUndefined(); // Not in cache yet
+
+				// Fetch explicitly
+				const fetchedItem = await mockResource.get?.item(123);
+				expect(fetchedItem).toBeDefined();
+
+				// Hook would trigger resolver (not tested here as it needs React)
+				expect(() => mockResource.use?.item(123)).not.toThrow();
+			});
+
+			it('should support write path: mutate → invalidate → prefetch', async () => {
+				// Create new item
+				await mockResource.mutate?.create({
+					title: 'New',
+					status: 'active',
+				});
+
+				// Invalidate affected caches
+				mockResource.cache.invalidate.list();
+				mockResource.cache.invalidate.item(3);
+
+				// Prefetch updated data
+				await mockResource.cache.prefetch.list({ status: 'active' });
+
+				expect(mockResource.mutate?.create).toHaveBeenCalled();
+				expect(mockResource.cache.invalidate.list).toHaveBeenCalled();
+				expect(mockResource.cache.prefetch.list).toHaveBeenCalled();
+			});
+
+			it('should support complete CRUD flow with cache management', async () => {
+				// 1. Fetch initial data
+				const items = await mockResource.get?.list();
+				expect(items).toBeDefined();
+
+				// 2. Create new item
+				const newItem = await mockResource.mutate?.create({
+					title: 'New',
+					status: 'active',
+				});
+				expect(newItem?.id).toBe(3);
+
+				// 3. Invalidate lists
+				mockResource.cache.invalidate.list();
+
+				// 4. Update the item
+				await mockResource.mutate?.update(3, { title: 'Updated' });
+
+				// 5. Invalidate the item
+				mockResource.cache.invalidate.item(3);
+
+				// 6. Prefetch fresh data
+				await mockResource.cache.prefetch.item(3);
+
+				// 7. Delete the item
+				await mockResource.mutate?.remove(3);
+
+				// 8. Invalidate everything
+				mockResource.cache.invalidate.all();
+
+				// Verify all operations were called
+				expect(mockResource.get?.list).toHaveBeenCalled();
+				expect(mockResource.mutate?.create).toHaveBeenCalled();
+				expect(mockResource.mutate?.update).toHaveBeenCalled();
+				expect(mockResource.mutate?.remove).toHaveBeenCalled();
+				expect(mockResource.cache.invalidate.item).toHaveBeenCalled();
+				expect(mockResource.cache.invalidate.list).toHaveBeenCalled();
+				expect(mockResource.cache.invalidate.all).toHaveBeenCalled();
+			});
+
+			it('should provide event names for observers', () => {
+				// Event names can be used for tracking/logging
+				const eventLog: Array<{ event: string; data: { id: number } }> =
+					[];
+
+				// Simulate creating an item with event emission
+				eventLog.push({
+					event: mockResource.events?.created || '',
+					data: { id: 3 },
+				});
+
+				// Simulate updating an item
+				eventLog.push({
+					event: mockResource.events?.updated || '',
+					data: { id: 3 },
+				});
+
+				// Simulate removing an item
+				eventLog.push({
+					event: mockResource.events?.removed || '',
+					data: { id: 3 },
+				});
+
+				expect(eventLog).toHaveLength(3);
+				expect(eventLog[0]?.event).toBe('wpk.thing.created');
+				expect(eventLog[1]?.event).toBe('wpk.thing.updated');
+				expect(eventLog[2]?.event).toBe('wpk.thing.removed');
+			});
+		});
+
+		describe('grouped API vs thin-flat API equivalence', () => {
+			it('get.item should be equivalent to fetch', async () => {
+				// Both should call the same underlying method
+				expect(mockResource.get?.item).toBeDefined();
+				expect(mockResource.fetch).toBeDefined();
+			});
+
+			it('get.list should be equivalent to fetchList', async () => {
+				expect(mockResource.get?.list).toBeDefined();
+				expect(mockResource.fetchList).toBeDefined();
+			});
+			it('use.item should be equivalent to useGet', () => {
+				expect(mockResource.use?.item).toBeDefined();
+				expect(mockResource.useGet).toBeDefined();
+			});
+
+			it('use.list should be equivalent to useList', () => {
+				expect(mockResource.use?.list).toBeDefined();
+				expect(mockResource.useList).toBeDefined();
+			});
+
+			it('mutate.create should be equivalent to create', () => {
+				expect(mockResource.mutate?.create).toBeDefined();
+				expect(mockResource.create).toBeDefined();
+			});
+
+			it('mutate.update should be equivalent to update', () => {
+				expect(mockResource.mutate?.update).toBeDefined();
+				expect(mockResource.update).toBeDefined();
+			});
+
+			it('mutate.remove should be equivalent to remove', () => {
+				expect(mockResource.mutate?.remove).toBeDefined();
+				expect(mockResource.remove).toBeDefined();
+			});
+
+			it('cache.prefetch.item should be equivalent to prefetchGet', () => {
+				expect(mockResource.cache.prefetch.item).toBeDefined();
+				expect(mockResource.prefetchGet).toBeDefined();
+			});
+
+			it('cache.prefetch.list should be equivalent to prefetchList', () => {
+				expect(mockResource.cache.prefetch.list).toBeDefined();
+				expect(mockResource.prefetchList).toBeDefined();
+			});
+
+			it('cache.invalidate methods should wrap invalidate', () => {
+				expect(mockResource.cache.invalidate.item).toBeDefined();
+				expect(mockResource.cache.invalidate.list).toBeDefined();
+				expect(mockResource.cache.invalidate.all).toBeDefined();
+				expect(mockResource.invalidate).toBeDefined();
 			});
 		});
 	});
