+++ packages/kernel/src/resource/define.ts
@@ -6,339 +6,64 @@
  *
  * @see Product Specification § 4.1 Resources
  */
-import type * as WPData from '@wordpress/data';
-import { KernelError } from '../error/index.js';
-import {
-	interpolatePath,
-	registerStoreKey,
-	invalidate as globalInvalidate,
-} from './cache.js';
+import { KernelError } from '../error/KernelError.js';
+import { registerStoreKey, invalidate as globalInvalidate } from './cache.js';
 import { createStore } from './store.js';
-import { fetch as transportFetch } from '../http/fetch.js';
-import type {
-	ResourceConfig,
-	ResourceObject,
-	ResourceClient,
-	ListResponse,
-	CacheKeys,
-} from './types.js';
-
-type WPGlobal = { wp?: { data?: typeof WPData } };
-
-/**
- * Default cache key generators
- *
- * Used when custom cacheKeys are not provided in config
- * @param resourceName
- */
-function createDefaultCacheKeys(resourceName: string): Required<CacheKeys> {
-	return {
-		list: (query) => [resourceName, 'list', JSON.stringify(query || {})],
-		get: (id) => [resourceName, 'get', id],
-		create: (data) => [resourceName, 'create', JSON.stringify(data || {})],
-		update: (id) => [resourceName, 'update', id],
-		remove: (id) => [resourceName, 'remove', id],
-	};
-}
-
-/**
- * Validate resource configuration
- *
- * Throws DeveloperError for invalid configs to catch issues at dev time
- *
- * @param config - Resource configuration to validate
- * @throws DeveloperError if configuration is invalid
- */
-function validateConfig<T, TQuery>(config: ResourceConfig<T, TQuery>): void {
-	// Validate name
-	if (!config.name || typeof config.name !== 'string') {
-		throw new KernelError('DeveloperError', {
-			message: 'Resource config must have a valid "name" property',
-			data: {
-				validationErrors: [
-					{
-						field: 'name',
-						message: 'Required string property',
-					},
-				],
-			},
-		});
-	}
-
-	if (!/^[a-z][a-z0-9-]*$/.test(config.name)) {
-		throw new KernelError('DeveloperError', {
-			message: `Resource name "${config.name}" must be lowercase with hyphens only (kebab-case)`,
-			data: {
-				validationErrors: [
-					{
-						field: 'name',
-						message:
-							'Must match pattern: lowercase letters, numbers, hyphens',
-					},
-				],
-			},
-		});
-	}
-
-	// Validate routes
-	if (!config.routes || typeof config.routes !== 'object') {
-		throw new KernelError('DeveloperError', {
-			message: 'Resource config must have a "routes" object',
-			data: {
-				validationErrors: [
-					{
-						field: 'routes',
-						message: 'Required object property',
-					},
-				],
-			},
-			context: { resourceName: config.name },
-		});
-	}
-
-	// At least one route must be defined
-	const routeKeys = Object.keys(config.routes);
-	if (routeKeys.length === 0) {
-		throw new KernelError('DeveloperError', {
-			message: `Resource "${config.name}" must define at least one route`,
-			data: {
-				validationErrors: [
-					{
-						field: 'routes',
-						message:
-							'At least one route (list, get, create, update, remove) required',
-					},
-				],
-			},
-			context: { resourceName: config.name },
-		});
-	}
-
-	// Validate each route definition
-	const validRouteNames = ['list', 'get', 'create', 'update', 'remove'];
-	for (const [routeName, route] of Object.entries(config.routes)) {
-		if (!validRouteNames.includes(routeName)) {
-			throw new KernelError('DeveloperError', {
-				message: `Invalid route name "${routeName}" in resource "${config.name}"`,
-				data: {
-					validationErrors: [
-						{
-							field: `routes.${routeName}`,
-							message: `Must be one of: ${validRouteNames.join(', ')}`,
-							code: 'INVALID_ROUTE_NAME',
-						},
-					],
-				},
-				context: { resourceName: config.name },
-			});
-		}
-
-		if (!route || typeof route !== 'object') {
-			throw new KernelError('DeveloperError', {
-				message: `Route "${routeName}" in resource "${config.name}" must be an object`,
-				data: {
-					validationErrors: [
-						{
-							field: `routes.${routeName}`,
-							message: 'Must be an object with path and method',
-							code: 'INVALID_ROUTE_TYPE',
-						},
-					],
-				},
-				context: { resourceName: config.name },
-			});
-		}
-
-		if (!route.path || typeof route.path !== 'string') {
-			throw new KernelError('DeveloperError', {
-				message: `Route "${routeName}" in resource "${config.name}" must have a valid "path"`,
-				data: {
-					validationErrors: [
-						{
-							field: `routes.${routeName}.path`,
-							message: 'Required string property',
-							code: 'MISSING_PATH',
-						},
-					],
-				},
-				context: { resourceName: config.name },
-			});
-		}
-
-		if (!route.method || typeof route.method !== 'string') {
-			throw new KernelError('DeveloperError', {
-				message: `Route "${routeName}" in resource "${config.name}" must have a valid "method"`,
-				data: {
-					validationErrors: [
-						{
-							field: `routes.${routeName}.method`,
-							message:
-								'Required string property (GET, POST, PUT, PATCH, DELETE)',
-							code: 'MISSING_METHOD',
-						},
-					],
-				},
-				context: { resourceName: config.name },
-			});
-		}
-
-		const validMethods = ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'];
-		if (!validMethods.includes(route.method)) {
-			throw new KernelError('DeveloperError', {
-				message: `Invalid HTTP method "${route.method}" for route "${routeName}" in resource "${config.name}"`,
-				data: {
-					validationErrors: [
-						{
-							field: `routes.${routeName}.method`,
-							message: `Must be one of: ${validMethods.join(', ')}`,
-							code: 'INVALID_METHOD',
-						},
-					],
-				},
-				context: { resourceName: config.name },
-			});
-		}
-	}
-}
-
-/**
- * Create resource client
- *
- * Generates typed REST client methods based on configured routes.
- *
- * @param config - Validated resource configuration
- * @return Resource client with typed methods
- */
-function createClient<T, TQuery>(
-	config: ResourceConfig<T, TQuery>
-): ResourceClient<T, TQuery> {
-	const client: ResourceClient<T, TQuery> = {};
-
-	// List method
-	if (config.routes.list) {
-		client.list = async (query?: TQuery): Promise<ListResponse<T>> => {
-			const response = await transportFetch<{
-				items?: T[];
-				total?: number;
-				hasMore?: boolean;
-				nextCursor?: string;
-			}>({
-				path: config.routes.list!.path,
-				method: 'GET',
-				query: query as Record<string, unknown>,
-			});
-
-			// Normalize response to ListResponse format
-			// Support both array responses and object responses with items property
-			const items = Array.isArray(response.data)
-				? response.data
-				: response.data.items || [];
-
-			return {
-				items,
-				total: response.data.total,
-				hasMore: response.data.hasMore,
-				nextCursor: response.data.nextCursor,
-			};
-		};
-	}
-
-	// Get method
-	if (config.routes.get) {
-		client.get = async (id: string | number): Promise<T> => {
-			const path = interpolatePath(config.routes.get!.path, { id });
-
-			const response = await transportFetch<T>({
-				path,
-				method: 'GET',
-			});
-
-			return response.data;
-		};
-	}
-
-	// Create method
-	if (config.routes.create) {
-		client.create = async (data: Partial<T>): Promise<T> => {
-			const response = await transportFetch<T>({
-				path: config.routes.create!.path,
-				method: 'POST',
-				data,
-			});
-
-			return response.data;
-		};
-	}
-
-	// Update method
-	if (config.routes.update) {
-		client.update = async (
-			id: string | number,
-			data: Partial<T>
-		): Promise<T> => {
-			const path = interpolatePath(config.routes.update!.path, { id });
-
-			const response = await transportFetch<T>({
-				path,
-				method: config.routes.update!.method as 'PUT' | 'PATCH',
-				data,
-			});
-
-			return response.data;
-		};
-	}
-
-	// Remove method
-	if (config.routes.remove) {
-		client.remove = async (id: string | number): Promise<void> => {
-			const path = interpolatePath(config.routes.remove!.path, { id });
-
-			await transportFetch<void>({
-				path,
-				method: 'DELETE',
-			});
-		};
-	}
-
-	return client;
-}
+import { validateConfig } from './validation.js';
+import { createClient } from './client.js';
+import { createDefaultCacheKeys, type WPGlobal } from './utils.js';
+import {
+	createSelectGetter,
+	createUseGetter,
+	createGetGetter,
+	createMutateGetter,
+	createCacheGetter,
+	createStoreApiGetter,
+	createEventsGetter,
+} from './grouped-api.js';
+import type { CacheKeys, ResourceConfig, ResourceObject } from './types.js';
 
 /**
  * Define a resource with typed REST client
  *
  * Creates a resource object with:
- * - Typed client methods (list, get, create, update, remove)
+ * - Typed client methods (fetchList, fetch, create, update, remove)
  * - Store key for @wordpress/data registration
  * - Cache key generators for invalidation
  * - Route definitions
+ * - Thin-flat API (useGet, useList, prefetchGet, prefetchList, invalidate, key)
+ * - Grouped API (select.*, use.*, get.*, mutate.*, cache.*, storeApi.*, events.*)
  *
- * @template T - Resource entity type (e.g., Thing)
- * @template TQuery - Query parameters type for list operations (e.g., { q?: string })
+ * @template T - Resource entity type (e.g., TestimonialPost)
+ * @template TQuery - Query parameters type for list operations (e.g., { search?: string })
  * @param    config - Resource configuration
  * @return Resource object with client methods and metadata
  * @throws DeveloperError if configuration is invalid
  * @example
  * ```ts
- * const thing = defineResource<Thing, { q?: string }>({
- *   name: 'thing',
+ * const testimonial = defineResource<TestimonialPost, { search?: string }>({
+ *   name: 'testimonial',
  *   routes: {
- *     list: { path: '/wpk/v1/things', method: 'GET' },
- *     get: { path: '/wpk/v1/things/:id', method: 'GET' },
- *     create: { path: '/wpk/v1/things', method: 'POST' }
+ *     list: { path: '/wpk/v1/testimonials', method: 'GET' },
+ *     get: { path: '/wpk/v1/testimonials/:id', method: 'GET' },
+ *     create: { path: '/wpk/v1/testimonials', method: 'POST' }
  *   },
  *   cacheKeys: {
- *     list: (q) => ['thing', 'list', q?.q],
- *     get: (id) => ['thing', 'get', id]
+ *     list: (q) => ['testimonial', 'list', q?.search],
+ *     get: (id) => ['testimonial', 'get', id]
  *   }
  * });
  *
- * // Use client methods
- * const items = await thing.list({ q: 'search' });
- * const item = await thing.get(123);
+ * // Thin-flat API
+ * const { items } = await testimonial.fetchList({ search: 'excellent' });
+ * const item = await testimonial.fetch(123);
+ * testimonial.invalidate([['testimonial', 'list']]);
  *
- * // Use metadata
- * console.log(thing.storeKey); // 'wpk/thing'
- * invalidate(thing.cacheKeys.list({ q: 'search' }));
+ * // Grouped API
+ * const cached = testimonial.select.item(123);
+ * await testimonial.get.item(123); // Always fresh from server
+ * await testimonial.mutate.create({ title: 'Amazing!' });
+ * testimonial.cache.invalidate.all();
  * ```
  */
 export function defineResource<T = unknown, TQuery = unknown>(
@@ -611,6 +336,46 @@ export function defineResource<T = unknown, TQuery = unknown>(
 				(v): v is string | number | boolean =>
 					v !== null && v !== undefined
 			);
+		},
+
+		// Grouped API: Pure selectors (no fetching)
+		// ===================================================================
+		// GROUPED API - Organized namespaces for power users
+		// ===================================================================
+
+		// Grouped API: Pure selectors (read from cache)
+		get select() {
+			return createSelectGetter<T, TQuery>(config).call(this);
+		},
+
+		// Grouped API: React hooks (alias to thin-flat)
+		get use() {
+			return createUseGetter<T, TQuery>().call(this);
+		},
+
+		// Grouped API: Explicit data fetching (bypass cache, direct REST calls)
+		get get() {
+			return createGetGetter<T, TQuery>(config).call(this);
+		},
+
+		// Grouped API: Mutations
+		get mutate() {
+			return createMutateGetter<T, TQuery>(config).call(this);
+		},
+
+		// Grouped API: Cache control
+		get cache() {
+			return createCacheGetter<T, TQuery>(config, cacheKeys).call(this);
+		},
+
+		// Grouped API: Store access
+		get storeApi() {
+			return createStoreApiGetter<T, TQuery>().call(this);
+		},
+
+		// Grouped API: Event names
+		get events() {
+			return createEventsGetter<T, TQuery>(config).call(this);
 		},
 	};
